msm8998-common: rootdir: Directly exec init.qcom.usb.sh helper
Directly execute the init.qcom.usb.sh helper in the 'on boot'
and 'on charger' actions rather than spawning it as a oneshot
service. After that, directly update sys.usb.config with the
from persist.vendor.usb.config. This streamlines and guarantees
it will execute synchronously which can help avoid rare race
conditions with other USB property triggers that could execute
in parallel.

Change-Id: Iaf798b146d071b11857a4cf54fbaf7ae0d5d2431