msm8998-common: gps: Apgs subscriber mananger fix
Currently the subscribers are created such that they are
not waiting for ConnectivityManager close event before
they get removed and thereby notify modem of the state
change. This may cause issue if modem immediately issues
a next ATL request. The state machine in AFW, while it is
in the middle tearing down a connection, is not able to
handle a next bringup request gracefully. Now we change
all the subscribers to have to wait for the tear down msg
from AFW before we remove the subscriber and notify modem
of the complete of the operation.

Change-Id: I05624b1cbbbc71eb189f9a0d73869d359b96cecd
CRs-Fixed: 2917823